change Sigs type
map from state to signature, not hash to signature

The core reason why the hash approach does not work well in all cases: we cannot always predict which states we expect.

Especially in a ledger channel with a hub, there may be many virtual channels, with the respective guarantees  jostling for position in the outcome. The ordering isn't terribly important but will effect the hash.