feat!: simplify state encoding logic
this removes the "nested" nature of the state encoding logic.

previously, the outcome and app data were encoded and hashed,
which were then fit into a fixed size struct containing a mix of
bytes32, uint256 and bool values.

Now, all state attributes are encoded into a single dynamically sized
struct. This makes the code easier to understand, but more importantly,
reduces the amount of cryptography that happens both on-chain and
off-chain, even in the happy path.