do not encode outcome separately
Before this change, we expect a solidity contract to define a struct

```
struct state {
  // ...
  outcome bytes
}
```

This means that to verify signatures of a state, we first need to encode
its outcome as bytes, plug it into a state struct, and then encode that
state as bytes.

The change in this PR allows a solidity contract to define a state which
has a property outcome that is itself a struct.
```
struct state {
  // ...
  outcome Outcome
}
```
As such, to encode a state, we would no longer need to encode outcome
first.

This is a more elegant and efficient approach which should save some
gas: see #133 (comment).