add directionality to chans connecting engine and chain service
Directionality is a great feature which I think we want. It provides greater type safety at compile time.

A subtlety is that we would like a given channel to be seen by the engine as a send and by the chain service as a recieve. I don't know of a way to do that.

The best I came up with is for the chain service to create two *bidirectional* private chans, and expose them through getters which narrow the type for consumers (i.e. the engine). So we don't have much safety in the chain service itself (we could still try to send on the channel we are supposed to be recieving on), outside of this package there is good type safety (the engine knows exactly what to do with each channel)