replace direct-fund.signAndStore with channel.SignAndAdd
This improves the centralization of code -- we will want to share this module with other protocols.

It also narrows the scope of the function somewhat, to only handling setup states.

It also replaces some errors with panics, along code paths that we should never hit (unless there is a severe bug).

We will panic if we have already signed the state.