use participants in V as source of truth
when constructing expected guarantees and ledger requests.

Currently, the expected guarantees as well as the ledger request side effects crafted by the virtualfund protocol are not always correct. This is because they are crafted using a relative "me/them" approach. As per the SATP paper (which is the specification for this protocol), the guarantees need to be crafted in absolute terms, depending on the index (here called myRole) of each client in the participant list of the virtual channel.