TestMessageService: remove pairwise connections
Instead of connecting peers in a pairwise fashion, this now defines a
Broker struct, which is simply a address -> incoming message channel
lookup.

(For some reason, the incoming message channel is called `out` -- why?)

This makes it easier to create more dynamic test scenarios: each
message service is already configured to receive messages when
NewTestMessageService returns.