gas benchmarking fixtures use a trivialapp
Previously, we could skip the application check altogether by having all participants countersign the same state.

But now, we must check the application. Hence there is a call to a contract, which previously did not need to be deployed, but now needs to be.

For now we will use the TrivialApp. Longer term we would want to use non trivial apps depending on the channel type.

We use some tricks to predict the contract address, since we need to know this _before_ the test runner starts.