remove test
This test is designed to compute a benchmark and print it to the screen. However, unless tests are run on CI with the -verbose flag (they currently aren't), we won't see the benchmark value printed.

Given the choice of getting that value persisted on CI, versus deleting the test in favour of benchmarking via our testground repo, it is much easier to do the former!