do not serialize empty structs in a messagePayload
as the messagePayload type is used, it either has a non-empty
SignedState or a non-empty SignedProposal.

This commit defines a custom JSON marshaler for messagePayload,
only serializing the struct that should be present. This reduces the
size of the message as well as increases readability to developers
in debug sessions.

It is good to solve the problem now, as will be exacerbated as we
add more values to the message format.