Add ConsensusApp.sol (#754)
* update typeconversions

* update challenge API

* add ConsensusApp.sol

* add ConsensusApp bindings

* introduce consensusapp package

* update challenge_test.go

* add test for ConsensusApp

NOTE there is one failing test which I have skipped. It seems unresolved exactly where to do signature validation? If inside the app definition, why provide the signedBy bitmask? If outside, why supply the sigs? And we need to put that validation somewhere.

* consensus app uses short cutting

unskip test


* deploy contacts from different EOAs

I'm not sure why this was necessary. perhaps something to do with account nonces?

* Update generate-adjudicator-bindings.sh

* npm install

* streamline script