chainservice returns errors in SendTransaction instead of panicking
go-eth transaction errors fall into 2 categories:
1. An error occurs when transaction is constructed and injected into the
pending pool.
2. An error occurs when transaction is mined.

If an error of the first category is encountered, chainservice returns the
error. The reasoning is that this type of error is likely caused by
an incorrect chainservice configuration or a reference to a contract
that is not deployed. A retry is unlikely to resolve these issues. The
error could be caused by an issue with a go-eth backend. We should explore
if and how often that occurs.

An error from the second category merits a retry before the chainservice
gives on the transaction.