implement Payment/Receipt managers
This includes a minimal implementation of a payment and receipt manager,
making the assumption that some other component is responsible for
sending vouchers across the wire.

A voucher is a simple struct containing a channel ID and an amount. This
voucher is signed by the sender in a virtual channel, and sent to the
receiver in the channel.

If the sender or hub do not cooperatively close the virtual channel, any
voucher signed by the sender can be redeemed by the receiver through a
forced transition on-chain, pending application rules implementation.