Remove fatalError and use fatalF instead
fatalError does not provide context for where in the codebase the
error occurred. It does print out the error that caused the panic, which
is useful. But we then see panics like:

FATAL ERROR
read tcp 192.18.0.9:58114->192.18.0.5:8545: i/o timeout

From logs like this, it is unclear what call triggered the panic.

With fatalF, more context can be supplied.