listeners do nothing in infinite loop until c.Conn is populated
I'm not massively happy with the way this works. But it shows one major bug we had:
we need to inspect the nullity of the underlying (embedded) websocket connection, not the (embedding) webSocketConnection type.

This change unblocks the test a bit -- we can get all the way to defunding before the disabled handlers (previous commit) cause things to stall out