change enforcement mechanism...
Previously, the candidate outcome was inspected to ensure the
challenger hasn't awarded themselves 'too much'.

BUT: This check did nothing to prevent the complete erasure of
the other party's funds in the channel.

With this change, the candidate outcome is inspected to ensure that
'too much' isn't deducted from the counterparty.