Use a mutex and single chan to implement engine events
The mutex l is used to synchronize the creation of the "broadcast chan". Since it is stored in a mapping keyed by objective id, this cannot be done ahead of time and must be created on demand by the first event subscriber. It should be possible to subscribe to (or wait on) an "objective completed" event before the objective has even been spawned -- this rules out make-ing the broadcast channel when the objective is created.

The entire mapping of chans is locked by the mutex.

The broadcast channel  c can have arbitrarily many subscribers. Those subscribers will all be unblocked when the broadcast channel is closed.

A subscriber will never miss the event, since they will unblock immediately if the broadcast channel is closed. If the channel doesn't exist yet, it will be created.