Use a `safesync.Map` of "broadcast chans" to emit "objective completed" events (#1170)
* Use a mutex and single chan to implement engine events

The mutex l is used to synchronize the creation of the "broadcast chan". Since it is stored in a mapping keyed by objective id, this cannot be done ahead of time and must be created on demand by the first event subscriber. It should be possible to subscribe to (or wait on) an "objective completed" event before the objective has even been spawned -- this rules out make-ing the broadcast channel when the objective is created.

The entire mapping of chans is locked by the mutex.

The broadcast channel  c can have arbitrarily many subscribers. Those subscribers will all be unblocked when the broadcast channel is closed.

A subscriber will never miss the event, since they will unblock immediately if the broadcast channel is closed. If the channel doesn't exist yet, it will be created.

* Update waitTimeForCompletedObjectiveIds to use new event API.

This helper is only complicated because:
- it races the event against a timeout
- it handles multiple objectiveIds

A much simpler thing to do (if the above concerns do not apply) is simply <-client.ObjectiveCompleteChan(id) .

* Copy over go-nitro Client Objective Completed event model to RPC client

see previous commits for explanation

* remove helper function

We aren't currently using the "advanced" functionality.

It's nicer to use the client API directly in the test if possible. It may be very slightly slower this way (or it may not be)

* rename / reorg data in rpc client

to match main go-nitro client better

* reduce usage of complicated (ish) helper

* typo

* move method back and improve godoc

* renaming

* make completedObjectives type private

* add LoadOrStore method to safesync.Map

* replace map+mutex with safesync.Map

* simplify ObjectiveCompleteChan

no need to alias the channel and use a goroutine

* update imports

artifact of rebase

* add ADR

* enhance adr

* enhance adr