virtualdefund.Objective has a reference to the virtual channel
rather than duplicating that information. This means that things like finalization proofs can be read from the store by other tasks/objectives. This will be important when it comes to handling challenges.

Other changes follow from this one, generally reading / writing to V, ensuring V is "related" to the objective (so it is stored alongside). We can reuse many helpers on V so quite a bit of code can be ripped out.