Switch websockets from github.com/nhooyr/ to github.com/gorilla/websocket
The primary reason for the change is that the websocket rpc client sometimes
hangs on close. It is unclear whether this is due to an incorrect usage of
the websocket library or an issue with the library itself.

Gorilla websockets is a popular library that is ergonomic to use. Unfortunately, the
github repository is archived. However, it doesn't seem like these types of
libraries see many updates anyways. Gorilla websockets was last updated in 7/2022
(https://github.com/gorilla/websocket/commits/master) and nhooyr websockets
was last updated on 12/2022.