Use a mutex to protect event queue
Currently, we have a goroutine that updates adds to the event queue.
We have another goroutine that reads from the event queue. This change:
- Protects the event queue with a mutex.
- Refactors notification handling so that both new events and new blocks
trigger queue processing. This allows us to remove the simulated chain
sleep that was added to make it more likely that an event update arrives
before a new block update.